pulled out tavern Scene from the ThreeD jsx file now will be working on integrating the MAIN scene and then rendering just the skybox with its elements inside of that scene.
