+++ client/src/containers/ThreeD/process.md
@@ -1,0 +1,30 @@
+## Three Js Plan
+
+## What We Have Done:
+
+*   Currently we have isolated the view of the scene to a div in the page
+*   I have integrated the chat function from discord into the page itself and created an almost overhead display for it
+*   Spotify needs to be integrated in the Threed page
+*   Current DM information or PC information like a heads up display
+
+
+## For better & more efficient rendering
+*   rescaling current scene to half size and changing camera FOV? fulcrum?
+
+## Travel/ Rendering
+*   click events to travel between scenes? Done within the 3d environment? On click on doors/portals?
+
+*   larger scene that contains multiple scenes for users to visit
+
+*   collapisble HUD?
+
+
+## Scenes to render
+    *   dungeon
+    *   sewer
+    *   city
+    *   forest
+    *   mountain
+    *   water?
+
+
