+++ client/src/containers/ThreeD/tavScene/TavernFloor.png
