+++ client/src/containers/ThreeD/tavScene/TavernWall.png
